# Conflicts:
#	Monopoly/.vs/Monopoly/v16/.suo
#	Monopoly/Board.cs
#	Monopoly/obj/Debug/net5.0/Monopoly.csprojAssemblyReference.cache

+++ Monopoly/Tiles/Buildable.cs
@@ -9,7 +9,7 @@ namespace Monopoly
 {
     interface Buildable
     {
-        public double getPrice();
+        public int getPrice();
 
         public int getAmountOfHouses();
